+++ mitra_adapters/src/init.rs
@@ -55,13 +55,6 @@ pub fn initialize_app(
     for warning in config_warnings {
         log::warn!("{}", warning);
     };
-    #[cfg(all(feature = "native-tls", target_env = "musl"))]
-    {
-        let is_found = openssl_probe::try_init_ssl_cert_env_vars();
-        if !is_found {
-            log::error!("certificate store not found");
-        };
-    };
     config
 }