+++ mitra_models/src/database/connect.rs
@@ -1,39 +1,26 @@
 use std::path::Path;
 
 use deadpool_postgres::Pool;
+use rustls::{
+    client::ClientConfig as TlsClientConfig,
+    Certificate,
+    Error as RustlsError,
+    RootCertStore,
+};
+use rustls_pemfile::certs;
 use tokio_postgres::{
     config::{Config as DatabaseConfig},
     Client,
 };
-
-#[cfg(feature = "native-tls")]
-use {
-    openssl::{
-        error::{ErrorStack as OpenSslError},
-        ssl::{SslConnector, SslMethod},
-    },
-    postgres_openssl::MakeTlsConnector,
-};
-
-#[cfg(feature = "rustls-tls")]
-use {
-    rustls::{
-        client::ClientConfig as TlsClientConfig,
-        Certificate,
-        Error as RustlsError,
-        RootCertStore,
-    },
-    rustls_pemfile::certs,
-    tokio_postgres_rustls::MakeRustlsConnect,
-};
+use tokio_postgres_rustls::MakeRustlsConnect;
 
 #[derive(thiserror::Error, Debug)]
 pub enum DatabaseConnectionError {
     #[error(transparent)]
     CertificateError(#[from] std::io::Error),
 
-    #[error("{0}")]
-    TlsError(String),
+    #[error(transparent)]
+    TlsError(#[from] RustlsError),
 
     #[error(transparent)]
     PostgresError(#[from] tokio_postgres::Error),
@@ -42,32 +29,6 @@ pub enum DatabaseConnectionError {
     PoolError(#[from] deadpool::managed::BuildError<tokio_postgres::Error>),
 }
 
-#[cfg(feature = "native-tls")]
-impl From<OpenSslError> for DatabaseConnectionError {
-    fn from(error: OpenSslError) -> Self {
-        Self::TlsError(error.to_string())
-    }
-}
-
-#[cfg(feature = "rustls-tls")]
-impl From<RustlsError> for DatabaseConnectionError {
-    fn from(error: RustlsError) -> Self {
-        Self::TlsError(error.to_string())
-    }
-}
-
-#[cfg(feature = "native-tls")]
-fn create_tls_connector(
-    ca_file_path: &Path,
-) -> Result<MakeTlsConnector, DatabaseConnectionError> {
-    let mut builder = SslConnector::builder(SslMethod::tls())?;
-    log::info!("using TLS CA file: {}", ca_file_path.display());
-    builder.set_ca_file(ca_file_path)?;
-    let connector = MakeTlsConnector::new(builder.build());
-    Ok(connector)
-}
-
-#[cfg(feature = "rustls-tls")]
 fn create_tls_connector(
     ca_file_path: &Path,
 ) -> Result<MakeRustlsConnect, DatabaseConnectionError> {